+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/MentoringService.java
@@ -27,6 +27,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class MentoringService {
 
     private final ImageService imageService;
+    private final CertificateService certificateService;
 
     private final MentoringRepository mentoringRepository;
     private final CategoryRepository categoryRepository;
@@ -135,8 +136,7 @@ public class MentoringService {
 
         Image profileImage = saveProfileImage(dto.profileImage(), savedMentoring);
 
-        certificateMapping(dto, savedMentoring);
-
+        certificateService.certificateMapping(dto, savedMentoring);
         return MentoringResponse.from(savedMentoring, categoryTitles, profileImage);
     }
 
@@ -154,36 +154,5 @@ public class MentoringService {
             return null;
         }
         return imageService.uploadImageToS3(profileImageFile, "profile-image", ImageType.MENTORING_PROFILE, mentoring.getId());
-    }
-
-    private void certificateMapping(RegisterMentoringDto dto, Mentoring savedMentoring) {
-        List<CertificateInfo> certificateInfos = dto.certificateInfos();
-        List<MultipartFile> certificateImageFiles = dto.certificateImages();
-
-        if (validateCertificateRequestData(certificateInfos, certificateImageFiles)) {
-            saveAllCertificates(certificateInfos, certificateImageFiles, savedMentoring);
-        }
-    }
-
-    private boolean validateCertificateRequestData(List<CertificateInfo> certificateInfos, List<MultipartFile> certificateImageFiles) {
-        return (certificateInfos != null && certificateImageFiles != null)
-                && (certificateInfos.size() == certificateImageFiles.size());
-    }
-
-    private void saveAllCertificates(List<CertificateInfo> certificateInfos, List<MultipartFile> certificateImageFiles, Mentoring savedMentoring) {
-        for (int i = 0; i < certificateInfos.size(); i++) {
-            CertificateInfo certificateInfo = certificateInfos.get(i);
-            MultipartFile certificateImageFile = certificateImageFiles.get(i);
-
-            saveCertificate(certificateInfo, certificateImageFile, savedMentoring);
-        }
-    }
-
-    private void saveCertificate(CertificateInfo request, MultipartFile certificateImageFile, Mentoring mentoring) {
-        final Certificate certificate = new Certificate(request.type(), request.title(), mentoring);
-        final Certificate savedCertificate = certificateRepository.save(certificate);
-        Long certificateId = savedCertificate.getId();
-
-        imageService.uploadImageToS3(certificateImageFile, "certificate-image", ImageType.CERTIFICATE, certificateId);
     }
 }
